backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
supabase==2.3.0

# Redis
redis==5.0.1
hiredis==2.3.2

# ML Libraries
xgboost==2.0.3
catboost==1.2.2
lightgbm==4.3.0
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
scipy==1.12.0

# PyTorch (optional, for neural networks)
torch==2.1.2
torchvision==0.16.2

# Feature Engineering
shap==0.44.1

# OCR
pytesseract==0.3.10
easyocr==1.7.1
Pillow==10.2.0
opencv-python==4.9.0.80

# Fuzzy Matching
rapidfuzz==3.6.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Task Queue
celery==5.3.6
flower==2.0.1

# Scheduling
apscheduler==3.10.4

# Monitoring
sentry-sdk[fastapi]==1.40.0
mlflow==2.10.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
